+++ scripts/build-package.js
@@ -25,7 +25,27 @@ const srcDir = path.resolve(rootDir, 'src/components');
 // tests and examples.
 function compileComponents() {
   return execa.shell(
-    `babel "${rootDir}/src/components" --out-dir ${outputDir} --ignore "**/examples","**/__tests__" --config-file ${rootDir}/babel.config.js`,
+    `babel "${rootDir}/src/components" --out-dir ${outputDir} --ignore "**/examples","**/__tests__,syntax-highlighting" --config-file ${rootDir}/babel.config.js`,
+    {
+      cwd: srcDir,
+      stdio: 'inherit'
+    }
+  );
+}
+
+function compileHelpers() {
+  return execa.shell(
+    `babel "${rootDir}/src/helpers" --out-dir ${outputDir}/helpers --config-file ${rootDir}/babel.config.js`,
+    {
+      cwd: srcDir,
+      stdio: 'inherit'
+    }
+  );
+}
+
+function compilePlugins() {
+  return execa.shell(
+    `babel "${rootDir}/src/plugins" --out-dir ${outputDir}/plugins --config-file ${rootDir}/babel.config.js`,
     {
       cwd: srcDir,
       stdio: 'inherit'
@@ -35,7 +55,7 @@ function compileComponents() {
 
 // Copy other src files that we want in the package.
 function copySrcFiles() {
-  return cpy(['css/*.css', 'data/*', 'plugins/*', 'helpers/*'], outputDir, {
+  return cpy(['css/*.css'], outputDir, {
     cwd: path.join(rootDir, 'src'),
     parents: true
   });
@@ -60,6 +80,7 @@ function createPackageJson() {
   delete publishable.optionalDependencies;
   delete publishable['lint-staged'];
   delete publishable.jest;
+  delete publishable.browserslist;
   fs.writeFileSync(
     path.resolve(outputDir, 'package.json'),
     JSON.stringify(publishable, null, 2)
@@ -71,6 +92,8 @@ del(outputDir)
   .then(() =>
     Promise.all([
       compileComponents(),
+      compileHelpers(),
+      compilePlugins(),
       copyMetaItems(),
       copySrcFiles(),
       createPackageJson()