+++ src/components/book-image/__tests__/book-image-test-cases.js
@@ -4,9 +4,14 @@ const testCases = {};
 
 testCases.basic = {
   component: BookImage,
+  description: 'Basic'
+};
+
+testCases.bigger = {
+  component: BookImage,
   description: 'Basic',
   props: {
-    size: '60'
+    size: 120
   }
 };