+++ src/components/note/__tests__/note.test.js
@@ -74,4 +74,22 @@ describe('note', () => {
       expect(tree).toMatchSnapshot();
     });
   });
+
+  describe(testCases.new.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      testCase = testCases.new;
+      wrapper = renderer.create(
+        React.createElement(testCase.component, testCase.props)
+      );
+      tree = wrapper.toJSON();
+    });
+
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
 });