+++ src/components/video/__tests__/video.test.js
@@ -1,0 +1,61 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import { testCases, noRenderCases } from './video-test-cases.js';
+
+describe('video', () => {
+  describe(testCases.basic.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      window.matchMedia = jest.fn().mockImplementation(query => {
+        return {
+          matches: false,
+          media: query,
+          onchange: null,
+          addListener: jest.fn(),
+          removeListener: jest.fn()
+        };
+      });
+
+      testCase = testCases.basic;
+      wrapper = renderer.create(
+        React.createElement(testCase.component, testCase.props)
+      );
+      tree = wrapper.toJSON();
+    });
+
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
+
+  describe(noRenderCases.reducedMotion.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      window.matchMedia = jest.fn().mockImplementation(query => {
+        return {
+          matches: true,
+          media: query,
+          onchange: null,
+          addListener: jest.fn(),
+          removeListener: jest.fn()
+        };
+      });
+
+      testCase = noRenderCases.reducedMotion;
+      wrapper = renderer.create(
+        React.createElement(testCase.component, testCase.props)
+      );
+      tree = wrapper.toJSON();
+    });
+
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
+});